Merged revision(s) 7040-7051 from trunk: continue documentation of ZEncoding.pas, no functional change ........ in order of https://zeoslib.sourceforge.io/viewtopic.php?f=50&t=128947 do the same for the PG-OID lobs too ........ some conversions of array types ........ continue documentation, no functional change ........ merged revision 7045 from branches/8.0-patches: Some changes to make the proxy server service work as a service ........
........
add TTime and TDate if the compiler doesn't know it
........
ADO don't log non OleExceptions
........
just test if TestTablePrivileges can retrieve a result. Implementation of exact result-tests should happen in the TestDbcUnits only
........
Merged revision(s) 7050 from branches/8.0-patches:
add a better determination for the configuration file on Linux, add a define to trigger the use of the CThreads unit on Linux.
........

........


git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/branches/8.0-patches@7052 d681b350-854c-439c-a81d-c0ce20710203