Revision(en) 8083-8093 von branches/8.0-patches zusammengeführt: Revision(en) 8048, 8079 von trunk zusammengeführt: By Joe @ Whale: Fix dangling FResultSet reference after a connection loss See [#576] for more information. ........ add TZCustomUniqueElementBinarySearchList with BinSearch-algo ascending find of elements ........
........

........
Fix a compilation problem for Delphi 10.4 and older -> The TCertificate.PublicKey member was only introduced with Delphi 11. This makes Trust on first use impossible on older Delphi versions.
........
Make sure, the internal proxy driver compiles with Delphi XE7 again.
........
Add a test for infinity and NaN for PostgreSQL with 'emulate_prepared' enabled.
........
Add a change that allows to have port numbers in the connection URI even if the host name is empty. PostgreSQL needs this to connect to unix domain sockets with non standard port names in the socket name. See:
https://zeoslib.sourceforge.io/viewtopic.php?f=20&t=191710&p=259457#p259299
........
Add some more recommendations by iteh. See:
https://zeoslib.sourceforge.io/viewtopic.php?f=20&t=191710&p=259457#p259300
........
Only send a password to PostgreSQL if it is not empty.
........


git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/trunk@8094 d681b350-854c-439c-a81d-c0ce20710203