Merged revisions 8328-8341 from branches/8.0-patches: Fix a parsing problem that is cause by selecting the wrong tokenizer in patch #8318 ........ A possible fix for a bug introduced when changing the schema of UncachedGetSchemas ........ Changes that fix problems when accessing Unicode streams with FPC in the proxy driver. ........
........

........
Replace $ENDIF with $IFEND to make older Delphis happy.
........
Fix another syntax error that only happens with new Delphi versions.
........
This seems to fix a bug in the test suites that only happens on 64 bits Delphi. Maybe it has to do with the changes in [r8313]
........
Enable the use of negative blob type numbers - as suggested by Firebird and Interbase. See
https://zeoslib.sourceforge.io/viewtopic.php?t=250628
for more information.
........
Add another constructor that can be called by fpc. Otherwise we generate problems in the test suite under fpc.
........
Remove ShowDebugInfo from the Lazarus test suite project. Maybe this helps to speedup compilation.
........
Also remove "ShowCond".
........
Make sure that the differing constructors are only available on the compiler that uses them.
........
By aehimself: Fetching the next batch of records should NOT be done in .RecNo
........


git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/trunk@8342 d681b350-854c-439c-a81d-c0ce20710203